// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
      categories: {
        Row: {
          active: boolean
          created_at: string
          display_order: number
          id: string
          image: string | null
          name: string
          slug: string
          updated_at: string
        }
        Insert: {
          active?: boolean
          created_at?: string
          display_order?: number
          id?: string
          image?: string | null
          name: string
          slug: string
          updated_at?: string
        }
        Update: {
          active?: boolean
          created_at?: string
          display_order?: number
          id?: string
          image?: string | null
          name?: string
          slug?: string
          updated_at?: string
        }
        Relationships: []
      }
// ...
      products: {
        Row: {
          active: boolean
          bestseller: boolean
          category_id: string | null
          created_at: string
          description: string | null
          discount_percent: number
          featured: boolean
          id: string
          name: string
          new_arrival: boolean
          price: number
          sale_price: number | null
          sizes: string[]
          sku: string | null
          slug: string
          stock: number
          updated_at: string
        }
        Insert: {
          active?: boolean
          bestseller?: boolean
          category_id?: string | null
          created_at?: string
          description?: string | null
          discount_percent?: number
          featured?: boolean
          id?: string
          name: string
          new_arrival?: boolean
          price: number
          sale_price?: number | null
          sizes?: string[]
          sku?: string | null
          slug: string
          stock?: number
          updated_at?: string
        }
        Update: {
          active?: boolean
          bestseller?: boolean
          category_id?: string | null
          created_at?: string
          description?: string | null
          discount_percent?: number
          featured?: boolean
          id?: string
          name?: string
          new_arrival?: boolean
          price?: number
          sale_price?: number | null
          sizes?: string[]
          sku?: string | null
          slug?: string
          stock?: number
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "products_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}